지역 관련 API 수정 및 불필요한 RegionPageController 삭제

+++ src/main/java/com/busan/orora/region/mapper/RegionMapper.java
@@ -16,6 +16,5 @@ public interface RegionMapper {
     
     List<SearchSpotsByRegionDto> searchSpotsByRegion(Long regionId);
 
-    // <select id="searchSpotsByRegionIds" parameterType="list" resultType="com.busan.orora.region.dto.SearchSpotsByRegionDto">
-    List<SearchSpotsByRegionDto> searchSpotsByRegionIds(@Param("list") List<Integer> regionIds);
+    List<SearchSpotsByRegionDto> searchSpotsByRegionIds(@Param("list") List<Long> regionIds);
 }
